fix "no sound on resume" when direct media is enabled
The phone on which the resume softkey is pressed was being sent a start
media transmission msg with an incorrect address (i.e. the asterisk address
instead of the other device address).

From what I understand, there is probably a cleaner way to fix this
problem by interacting with asterisk differently. This solution feels
more like a hack.