re-add check on subchan->hold in cb_ast_answer
This can happens in the following scenario:
* you call another user, which doesn't answer
* you then press the voicemail hard key

That said, both the old and new behaviour are wrong; you should probably
not be able to do 2 outgoing "ringout" calls at the same time, i.e. it
should not be possible to place on hold a ringout channel. Also, with
direct media, it doesn't work.