Start the callid at the epoch when the phone is registering
the 7940 has a bug and does not add calls to it's history if the callid
is lower than it's last history entry.

Starting the callid at the epoch ensures that the callid will be higher
than the last callid in the history unless the user makes more than one
call a second.

The bug is still present and new calls will not be in the call history
when the integer overflows (in 2038)