add backward compatibility to device_cfg->voicemail
So that if the format of the voicemail option in sccp.conf was:

  voicemail = 1234

then it is the same as if it was

  voicemail = 1234@<line context>

Previously, you could only specify the number part of the voicemail and
the context of the voicemail was always considered to be the context of
your line. With these changes, you can now specify a context with your
voicemail, and it does not have to be in the same context as the line,
yet it does not break existing sccp.conf file.